add review feedback
remove mockito compile dependency from sample apps
add since java doc comment
use Spock Mock where possible
