+++ tests/spring-cloud-contract-stub-runner-amqp/src/main/java/org/springframework/cloud/contract/stubrunner/messaging/amqp/AmqpMessagingApplication.java
@@ -1,14 +1,9 @@
 package org.springframework.cloud.contract.stubrunner.messaging.amqp;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import static org.springframework.amqp.core.MessageProperties.CONTENT_TYPE_JSON;
-
-import java.util.concurrent.ExecutorService;
 
 import org.springframework.amqp.core.Exchange;
 import org.springframework.amqp.core.TopicExchange;
-import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter;
@@ -20,8 +15,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.Connection;
 
 @SpringBootApplication
 public class AmqpMessagingApplication {
@@ -55,20 +48,5 @@ public class AmqpMessagingApplication {
 	@Bean
 	public MessageListenerAdapter messageListenerAdapter(MessageSubscriber messageSubscriber, MessageConverter messageConverter) {
 		return new MessageListenerAdapter(messageSubscriber, messageConverter);
-	}
-
-	@Bean
-	public ConnectionFactory connectionFactory() {
-		com.rabbitmq.client.ConnectionFactory mockConnectionFactory = mock(com.rabbitmq.client.ConnectionFactory.class);
-		Connection mockConnection = mock(Connection.class);
-		Channel mockChannel = mock(Channel.class);
-		try {
-			when(mockConnectionFactory.newConnection((ExecutorService) null)).thenReturn(mockConnection);
-			when(mockConnection.isOpen()).thenReturn(true);
-			when(mockConnection.createChannel()).thenReturn(mockChannel);
-		} catch (Exception e) {
-			throw new RuntimeException(e);
-		}
-		return new CachingConnectionFactory(mockConnectionFactory);
 	}
 }
